backend/pyproject.toml
[tool.poetry]
name = "wikipulse-api"
version = "0.1.0"
description = "WikiPulse Service API"
authors = ["WikiPulse Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
aiokafka = "^0.10.0"
redis = {extras = ["asyncio"], version = "^5.0.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
boto3 = "^1.34.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
